fix: move functions and requests to ReservationsContext
