+++ src/components/Filter/index.tsx
@@ -4,27 +4,9 @@ import 'react-date-range/dist/styles.css';
 import 'react-date-range/dist/theme/default.css';
 import { DateRangePicker } from 'rsuite';
 import 'rsuite/dist/rsuite.css';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
+import { Controller, useForm } from 'react-hook-form';
 import { SelectItem } from '../SelectItem';
-import { api } from '../../services/api';
-
-interface IFilter {
-  selectHotel: string;
-  activityType: string;
-  dates: Date[];
-}
-
-interface IDates {
-  startDate: string;
-  endDate: string;
-}
-
-interface IReservetions {
-  id: number;
-  userId: number;
-  hotelId: number;
-  dates: IDates;
-}
+import { IFilter } from '../../contexts/ReservationsContext/types';
 
 export const Filter = () => {
   const {
@@ -34,7 +16,8 @@ export const Filter = () => {
     activityType,
     activityTypeChange,
     setHotels,
-    getAllHotels
+    getAllHotels,
+    submit
   } = useContext(ReservationsContext);
 
   const activitiesTypesArray = [
@@ -47,59 +30,6 @@ export const Filter = () => {
       name: 'aquatica'
     }
   ];
-
-  const reservedHotelsByDate = (date1Start: string, date1End: string, date2Start: string, date2End: string) => {
-
-    const start1 = new Date(date1Start);
-    const end1 = new Date(date1End);
-    const start2 = new Date(date2Start);
-    const end2 = new Date(date2End);
-
-   
-    return (start1 < end2) && (start2 < end1);
-    
-  };
-
-  const submit: SubmitHandler<IFilter> = async (data: IFilter)  => {
-
-    const {selectHotel, activityType, dates} = data;
-
-    console.log(data);
-
-    const token = localStorage.getItem('@TOKEN');
-
-    if(dates && selectHotel === ''){
-      try {
-        const reservedHotels = await api.get<IReservetions[]>('/reservedHotels', {
-          headers: {
-            Authorization: `Bearer ${token}`
-          }
-        });
-
-        const formatedDates = dates.map(date => date.toLocaleDateString());
-
-        const unavailableHotelsById = reservedHotels.data.map(reservation => {
-
-          const reservedDates = Object.values(reservation.dates);
-
-         if(reservedHotelsByDate(formatedDates[0], formatedDates[1], reservedDates[0], reservedDates[1])){
-          return reservation.hotelId;
-         }
-        });
-
-        if(unavailableHotelsById.length > 0){
-          const filteredHotels = hotels?.filter(hotel => !unavailableHotelsById.includes(hotel.id));
-          setHotels([...filteredHotels]); //  ver esse erro depois;
-        }
-      
-
-
-      } catch (error) {
-        console.log(error);
-      }
-    }
-
-  };
 
   const { register, handleSubmit, control } = useForm<IFilter>();
